Move types to playerTypes file rather than within Player file
