+++ features/players/components/__tests__/VideoControlIndicator.test.tsx
@@ -1,5 +1,4 @@
 import { render, screen } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
 import { VideoControlIndicator } from "../VideoControlIndicator";
 
 describe("Control indicator rendering and prop handling", () => {
